Module plonk::constraint_system::composer

The Composer is a Trait that is actually defining some kind of Circuit Builder for PLONK.

In that sense, here we have the implementation of the StandardComposer which has been designed in order to provide the maximum amount of performance while having a big scope in utility terms.

It allows us not only to build Add and Mul constraints but also to build ECC op. gates, Range checks, Logical gates (Bitwise ops) etc.

Structs

StandardComposer
A composer is a circuit builder and will dictate how a circuit is built We will have a default Composer called StandardComposer