Struct plonk::constraint_system::composer::StandardComposer

pub struct StandardComposer {
    pub zero_var: Variable,
    // some fields omitted
}

A composer is a circuit builder and will dictate how a circuit is built We will have a default Composer called StandardComposer

Fields
zero_var: Variable

A zero variable that is a part of the circuit description. We reserve a variable to be zero in the system This is so that when a gate only uses three wires, we set the fourth wire to be the variable that references zero

Methods

impl StandardComposer

pub fn preprocess(
&mut self,
commit_key: &ProverKey,
transcript: &mut dyn TranscriptProtocol,
domain: &EvaluationDomain
) -> PreProcessedCircuit

Computes the pre-processed polynomials So the verifier can verify a proof made using this circuit

pub fn prove(
&mut self,
commit_key: &ProverKey,
preprocessed_circuit: &PreProcessedCircuit,
transcript: &mut dyn TranscriptProtocol
) -> Proof

Prove will compute the pre-processed polynomials and produce a proof

pub fn circuit_size(&self) -> usize

Returns the number of gates in the circuit

impl StandardComposer

pub fn new() -> Self

Generates a new empty StandardComposer with all of it's fields set to hold an initial capacity of 0.

Warning

The usage of this may cause lots of re-allocations since the Composer holds Vec for every polynomial, and these will need to be re-allocated each time the circuit grows considerably.

pub fn split_tx_poly(
&self,
n: usize,
t_x: &Polynomial
) -> (Polynomial, Polynomial, Polynomial, Polynomial)

Split t(X) poly into 3 degree n polynomials.

pub fn with_expected_size(expected_size: usize) -> Self

Creates a new circuit with an expected circuit size. This will allow for less reallocations when building the circuit since the Vecs will already have an appropiate allocation at the beggining of the composing stage.

pub fn add_input(&mut self, s: Scalar) -> Variable

Add Input first calls the Permutation struct to generate and allocate a new variable var. The composer then links the Variable to the Scalar and returns the Variable for use in the system.

pub fn add_gate(
&mut self,
a: Variable,
b: Variable,
c: Variable,
q_l: Scalar,
q_r: Scalar,
q_o: Scalar,
q_c: Scalar,
pi: Scalar
) -> Variable

Adds a width-3 add gate to the circuit, linking the addition of the provided inputs, scaled by the selector coefficients with the output provided.

pub fn add(
&mut self,
q_l_a: (Scalar, Variable),
q_r_b: (Scalar, Variable),
q_c: Scalar,
pi: Scalar
) -> Variable

Adds a big_addition_gate with the left and right inputs and it's scaling factors, computing & returning the output (result) Variable, and adding the corresponding addition constraint.

This type of gate is usually used when we don't need to have the largest amount of performance as well as the minimum circuit-size possible. Since it defaults some of the selector coeffs = 0 in order to reduce the verbosity and complexity.

Forces q_l * w_l + q_r * w_r + q_c + PI = w_o(computed by the gate).

pub fn big_add(
&mut self,
q_l_a: (Scalar, Variable),
q_r_b: (Scalar, Variable),
q_4_d: (Scalar, Variable),
q_c: Scalar,
pi: Scalar
) -> Variable

Adds a big_addition_gate with the left, right and fourth inputs and it's scaling factors, computing & returning the output (result) Variable and adding the corresponding addition constraint.

This type of gate is usually used when we don't need to have the largets ammount of performance and the minimum circuit-size possible. Since it defaults some of the selector coeffs = 0 in order to reduce the verbosity and complexity.

Forces q_l * w_l + q_r * w_r + q_4 * w_4 + q_c + PI = w_o(computed by the gate).

pub fn big_add_gate(
&mut self,
a: Variable,
b: Variable,
c: Variable,
d: Variable,
q_l: Scalar,
q_r: Scalar,
q_o: Scalar,
q_4: Scalar,
q_c: Scalar,
pi: Scalar
) -> Variable

Adds a width-4 add gate to the circuit and it's corresponding constraint.

This type of gate is usually used when we need to have the largest amount of performance and the minimum circuit-size possible. Since it allows the end-user to set every selector coefficient as scaling value on the gate eq.

pub fn mul_gate(
&mut self,
a: Variable,
b: Variable,
c: Variable,
q_m: Scalar,
q_o: Scalar,
q_c: Scalar,
pi: Scalar
) -> Variable

Adds a width-3 add gate to the circuit linking the product of the provided inputs scaled by the selector coefficient q_m with the output provided scaled by q_o.

Note that this gate requires to provide the actual result of the gate (output wire) since it will just add a mul constraint to the circuit.

pub fn big_mul_gate(
&mut self,
a: Variable,
b: Variable,
c: Variable,
d: Variable,
q_m: Scalar,
q_o: Scalar,
q_c: Scalar,
q_4: Scalar,
pi: Scalar
) -> Variable

Adds a width-4 big_mul_gate with the left, right and fourth inputs and it's scaling factors, computing & returning the output (result) Variable and adding the corresponding mul constraint.

This type of gate is usually used when we need to have the largest amount of performance and the minimum circuit-size possible. Since it allows the end-user to setup all of the selector coefficients.

Forces q_l * (w_l + w_r) + w_4 * q_4 + + q_c + PI = q_o * w_o(computed by the gate).

pub fn mul(
&mut self,
q_m: Scalar,
a: Variable,
b: Variable,
q_c: Scalar,
pi: Scalar
) -> Variable

Adds a simple and basic addition to the circuit between to Variables returning the resulting Variable.

pub fn big_mul(
&mut self,
q_m: Scalar,
a: Variable,
b: Variable,
q_4_d: (Scalar, Variable),
q_c: Scalar,
pi: Scalar
) -> Variable

Adds a width-4 big_mul_gate with the left, right and fourth inputs and it's scaling factors, computing & returning the output (result) Variable and adding the corresponding mul constraint.

This type of gate is usually used when we don't need to have the largest ammount of performance and the minimum circuit-size possible. Since it defaults some of the selector coeffs = 0 in order to reduce the verbosity and complexity.

Forces q_l * (w_l + w_r) + w_4 * q_4 + q_c + PI = w_o(computed by the gate).

pub fn poly_gate(
&mut self,
a: Variable,
b: Variable,
c: Variable,
q_m: Scalar,
q_l: Scalar,
q_r: Scalar,
q_o: Scalar,
q_c: Scalar,
pi: Scalar
) -> (Variable, Variable, Variable)

Adds a width-3 poly gate. This gate gives total freedom to the end user to implement the corresponding circuits in the most optimized way possible because the under has access to the whole set of variables, as well as selector coefficients that take part in the computation of the gate equation.

The final constraint added will force the following: (a * b) * q_m + a * q_l + b * q_r + q_c + PI + q_o * c = 0.

pub fn constrain_to_constant(
&mut self,
a: Variable,
constant: Scalar,
pi: Scalar
)

Adds a gate which is designed to constrain a Variable to have a specific constant value which is sent as a Scalar.

pub fn bool_gate(&mut self, a: Variable) -> Variable

Adds a boolean constraint (also known as binary constraint) where the gate eq. will enforce that the Variable received is either 0 or 1 by adding a constraint in the circuit.

Note that using this constraint with whatever Variable that is not representing a value equalling 0 or 1, will always force the equation to fail.

pub fn range_gate(&mut self, witness: Variable, num_bits: usize)

Adds a range-constraint gate that checks and constrains a Variable to be inside of the range [0,num_bits].

pub fn logic_xor_gate(
&mut self,
a: Variable,
b: Variable,
num_bits: usize
) -> Variable

Adds a logical XOR gate that performs the XOR between two values for the specified first num_bits returning a Variable holding the result.

Panics

If the num_bits specified in the fn params is odd.

pub fn logic_and_gate(
&mut self,
a: Variable,
b: Variable,
num_bits: usize
) -> Variable

Adds a logical AND gate that performs the bitwise AND between two values for the specified first num_bits returning a Variable holding the result.

Panics

If the num_bits specified in the fn params is odd.

pub fn assert_equal(&mut self, a: Variable, b: Variable)

Asserts that two variables are the same

pub fn add_dummy_constraints(&mut self)

This function is used to add a blinding factor to the witness polynomials