Module plonk::fft

FFT module contains the tools needed by the Composer backend to know and use the logic behind Polynomials. As well as the operations that the Composer needs to peform with them.

Structs

EvaluationDomain
Defines a domain over which finite field (I)FFTs can be performed. Works only for fields that have a large multiplicative subgroup of size that is a power-of-2.

Evaluations
Stores a polynomial in evaluation form.

Polynomial
Polynomial represents a polynomial in coeffiient form.