Struct plonk::fft::Polynomial

pub struct Polynomial {
    pub coeffs: Vec<Scalar>,
}

Polynomial represents a polynomial in coeffiient form.

Fields
coeffs: Vec<Scalar>

The coefficient of x^i is stored at location i in self.coeffs.

Methods

impl Polynomial

pub fn zero() -> Self

Returns the zero polynomial.

pub fn is_zero(&self) -> bool

Checks if the given polynomial is zero.

pub fn from_coefficients_slice(coeffs: &[Scalar]) -> Self

Constructs a new polynomial from a list of coefficients.

pub fn from_coefficients_vec(coeffs: Vec<Scalar>) -> Self

Constructs a new polynomial from a list of coefficients.

pub fn degree(&self) -> usize

Returns the degree of the polynomial.

pub fn evaluate(&self, point: &Scalar) -> Scalar

Evaluates self at the given point in the field.

pub fn rand<R: Rng>(d: usize, rng: &mut R) -> Self

Outputs a polynomial of degree d where each coefficient is sampled uniformly at random from the field F.

impl Polynomial

pub fn ruffini(&self, z: Scalar) -> Polynomial

Divides self by x-z using Ruffinis method