Module plonk::notes::prove_verify

This module contains the notes on how the prover algorithm is constructed for PLONK.

PLONK proof construction

Following on from the generic SNARK construction, here we will give the set up of a PLONK proof and show which steps need to be satisfied to utilise the protocol.

First we will explain the derivation and simplification of the arithmetic circuits.

PLONK uses both gate constraints and copy constraints, to collect like expressions. Using the same example of:

\[ \mathbf{W}_L \cdot \mathbf{W}_R = \mathbf{W}_O, \]

We can express multiples of the same wires in or out of the same gate, with the above equation.

Thus we have 'ith' gates, so the index from left or right across the circuit is mitigated for wires which are equal.

For example, in the two equations:

\[ \begin{aligned} \mathbf{A}_1 \circ X \cdot \mathbf{B}_1 \circ \X^{2} = \mathbf{C}_1\\ and \mathbf{A}_2 \circ X^{2} \cdot \mathbf{B}_2 \circ \X = \mathbf{C}_2\\ \end{aligned} \]

We can state the equalities that: \[ \begin{aligned} \mathbf{A}_1 = \mathbf{B}_2 \\ &\\ \mathbf{B}_1 = \mathbf{A}_2 \\ \end{aligned} \]

These are examples of constraints collected in PLONK. Which is done the same for addition gates, except the gate constrain satisfies:

\[ \mathbf{W}_L + \mathbf{W}_R = \mathbf{W}_O, \]

PLONK also uses 'copy constraints', which are used to associate wires, which have equality, from the entire circuit. These constriants are checked with a permutation argument. In essence, this checks that wires are not repeated by using randomness given by the verifier.

This process is better explained in the permutation notes.

After the constraints are made, they are formatted into a system of mumerical equations, which in PLONK are reduced to a small amount of polynomial equations which are capable of representing the constraints. PLONK allows us to combine the two gate equations by describing their relationship relative to the role in the circuit. PLONK also has constant, which are denoted as 'Q'. These values will change for each programme. The shape of the circuit is defined by these When they are combined with the gate equations, we get the polynomial equation for a reduced form as:

\[ \begin{aligned}
L = left \\ R = right \\ O = output \\ M = multiplication \\ C = constants \\

\mathbf{Q}_L_i \circ a_i + \mathbf{Q}_R_i \circ b_i + \mathbf{Q}_0_i \circ c_i + \mathbf{Q}_M_i \circ a_ib_i + \mathbf{Q}_C_i = 0 \end{aligned} \] This can be used for both addition and multiplication gates, where there values can be provided by the user depending on the circuit composition. For an addition gate, we derive it as follows:

\[ \begin{aligned} \mathbf{Q}_L_i = 1 \mathbf{Q}_R_i = 1 \mathbf{Q}_0_i = -1 \mathbf{Q}_M_i = 0 \mathbf{Q}_C_i = 1 \end{aligned} \]

Which results in: \[ \begin{aligned}
a_i + b_i - c_i = 0 \end{aligned} \]

For a multiplication gate, we derive it as follows:

\[ \begin{aligned} \mathbf{Q}_L_i = 0 \mathbf{Q}_R_i = 0 \mathbf{Q}_0_i = -1 \mathbf{Q}_M_i = 1 \mathbf{Q}_C_i = 0 \end{aligned} \]

Which results in: \[ \begin{aligned}
a_ib_i - c_i = 0 \end{aligned} \]

With this format, there is a specific method used to convert all the equations into polynomials form. Basically, to bundling these together, PLONK can take sets of equations and turn them into one single equation over polynomials. This called the evalutaion form. We are then able to use Lagrangian interpolation to convert to coefificient form. All this interpolation is doing is allowing us to evaluate a functions over specific points, for 'x' values, where the target polynomial is equal to '1' or '0'.

With these specific bases, we can derive the relation between all sets of equations into one single polynomial equation, where we have a vector of inputs to each gate type: \[ \begin{aligned} \mathbf{Q}_L(x) \circ a(x) + \mathbf{Q}_R(x) \circ b(x) + \mathbf{Q}_0(x) \circ c(x) + \mathbf{Q}_M(x) \circ a(x)b(x) + \mathbf{Q}_C(x) = 0 \end{aligned} \]

The utility for this in PLONK, as a univeral SNARK, is that any operation or relationship that holds with the inputted vectors, will also hold over the polynomial.

In order to check this in PLONK, a 'vanishing polynomial' is introduced. Which is just a polynomial equal to zero for all the points we evaluate at. So this means that the vectors will be divisible by this vanishing polynomial, if the expression we check does indeed hold.

To summarise what the PLONK proof must satisfy:

The generation of copy and gate constraints where the former are representative of permuted wires. Generate all of the polynomials to
be checked against the vanishing polynomials. Take all of the wire values and convert them into three polynomials, A(x), B(x), C(x). Check the polynomials at some random Z(z), by making commitments and checking the evaluation form. Then commit all evalution polynomials forms for the verfier.

Lagrangian polynomials

The use of polynomials in the PLONK proving scheme refers to specific evaluation domains, named Lagrangian polynomials,
based on interpolation of two functions of a particular group elements. The following section gives a more comprehensive understanding to the way in which these polynomials are formed, given certain inputs.

Langrangian polynomials are introduced as a means of constructing continous functions from discrete data. With alternative polynomial constructions, discrete data sets can be approximated; Langrangian polynomials, however, form a solution that fits data exactly. This is achieved through interpolation, which finds a linear combination of 'n' inputted functions with respect to a given data set which imposes 'n' constraints and computes an exact fitting solution.

Linear Algebra dictates that interpolation polynomial ought to be formed from the system {\mathbf{A}{x}} = {\mathbf{b}}, where b_i = y_i, i\ = 0,...,n\, and the entries of {\mathbf{A}} are defined by {\mathbf{a}}_{\operatorname{ij}} = {\mathbf{p}}{(x_i)},i\,j\ = 0,....,n, where x_0,x_1,...,x_n are the points at which data y_o, y_1,...,y_n are obtained, and {\mathbf{p_j}}{(x_i)} = x^{j}, j\ = 0,1,...,n. The basis {1,x\,..., x^{n}} of the space of polynomials degree n\+1 is valled the monomial basis, and the corresponding matrix A is called the Vandermode matrix for the points x_0, x_1,...,x_n.

Langrangian interpolation, however, has the matrix A, as the identity matrix. This stems from writing the interpolating polynomial as

\begin{aligned} {\mathbf{p_n }}{(x\)} = \sum_{j=0}^{n} y_i\mathcal{L}_n,j{(x)}, \end{aligned}

where the polynomials {\mathcal{L_n},_j(x_i)} have the property that \begin{aligned} {\mathcal{L_n},_j(x_i)} \begin{cases} 1 & \text{if $i$ = $j$}\ 0 & \text{if $i$ \neq $j$}\

\end{aligned}

The polynomials {\mathcal{L_n},_j}, j\ = 0,...,n, are intertpolations of the points x_0, x_1,...,x_n. Theya are commonly called the Lagrangian polynomials. They are wriiten in the form \begin{aligned} {\mathcal{L_{n,j}(x) = \prod_{k=0 k\neqj}^{n} frac{(x-x_{k}){(x_{j}-x_{k}) \end{aligned}

the unique solution polynomial of degree 'n' that satisfies this \begin{aligned} {\mathbf{p_n}}{(x_j)},i\,j\ = $f$(x_j), $j$ = 0,1,...1,n. \end{aligned}

This polynomial, {\mathbf{p_n}}{(x_j)} is called the interpolating polynomial of $f$(x).

To understand these as an expanded product argument, it can be written as

Given a set of k + 1 data points

data points

where no two
x_j are the same, the interpolation polynomial in the Lagrange form is a linear combination

Lagrange polynomial

of Lagrange basis polynomials Basis Polynomial

Example

To find the unique Lagranian interpolation of polynomial p_3(x), with degree \leq 3, from the following set of data points: \documentclass{article}

\begin{table} \begin{center} \label{Data points} \begin{tabular}{l|c|r} % <-- Alignments: 1st column left, 2nd middle and 3rd right, with vertical lines in between \textbf{Value 1} & \textbf{Value 2} & \textbf{Value 3}\ $i$ & $x_i$ & $y_i$ \ \hline 0 & -1 & 3\ 1 & 0 & -4\ 2 & 1 & 5\ 3 & 2 & -6 \end{tabular} \end{center} \end{table}

In order to find this polynomial, we must construct the Lagrangian of the form {\mathcal{L_{n,j}(x) = \prod_{i=0 i\neqj}^{n} frac{(x-x_{j}){(x_{j}-x_{i})

Calculated as follows:

$\mathcal{L}{3,0}(x) = \frac{(x-x{1})(x-x_{2})(x-x_{3})}{(x_{0}-x_{1})(x_{0}-x_{2})(x_{0}-x_{3})} = \frac{(x-0)(x-1)(x-2)}{(-1-0)(-1-1)(-1-2)} = \frac{x(x^{2}-3x+2)}{(-1)(-2)(-3)} = -\frac{1}{6}(x^{3}-3x^{2}+2x)$

$\mathcal{L}{3,1}(x) = \frac{(x-x{0})(x-x_{2})(x-x_{3})}{(x_{1}-x_{0})(x_{1}-x_{2})(x_{1}-x_{3})} = \frac{(x+1)(x-1)(x-2)}{(0+1)(0-1)(0-2)} = \frac{(x^2-1)(x-2)}{(1)(-1)(-2)} = \frac{1}{2}(x^{3}-2x^{2}-x+2)$

$\mathcal{L}{3,2}(x) = \frac{(x-x{0})(x-x_{1})(x-x_{3})}{(x_{2}-x_{0})(x_{2}-x_{1})(x_{2}-x_{3})} = \frac{(x+1)(x-0)(x-2)}{(1+1)(1-0)(1-2)} = \frac{x(x^2-x-2)}{(2)(1)(-1)} = -\frac{1}{2}(x^{3}-x^{2}-2x)$

$\mathcal{L}{3,3}(x) = \frac{(x-x{0})(x-x_{1})(x-x_{2})}{(x_{3}-x_{0})(x_{3}-x_{1})(x_{3}-x_{2})} = \frac{(x+1)(x-0)(x-1)}{(2+1)(2-0)(2-1)} = \frac{x(x^2-1)}{(3)(2)(1)} = \frac{1}{6}(x^{3}-x)$

This provides the unique polynomial which fits the given data sets, given by the interpolating polynomial

$p_{3}(x) = \sum_{j=0}^{3}y_{j}\mathcal{L}{3,j}(x) = y{0}\mathcal{L}{3,0}(x) + y{1}\mathcal{L}{3,1}(x) + y{2}\mathcal{L}{3,2}(x) + y{3}\mathcal{L}_{3,3}(x) = (3)(-\frac{1}{6})(x^{3}-3x^{2}+2x) + (-4)\frac{1}{2}(x^{3}-2x^{2}-x+2) + (5)(-\frac{1}{2})(x^{3}-x^{2}-2x) + (-6)\frac{1}{6}(x^{3}-x) = (-\frac{1}{2})(x^{3}-3x^{2}+2x) + (-2)(x^{3}-2x^{2}-x+2) + (-\frac{5}{2})(x^{3}-x^{2}-2x) - (x^{3}-x) = (-\frac{1}{2}-2-\frac{5}{2}-1)x^{3} + (\frac{3}{2} + 4 + \frac{5}{2})x^{2} + (-1+2+5+1)x - 4 = -6x^{3}+8x^{2}+7x-4$

Lagrangian polynomials are verifiable; if each x_{i}, for i\ \exist 0...n, is substituted into p_{n}(x), then we obtain p_{n}(x_{i}) = y_{i}.